Fix memory caching for lambda calls
sha256_hashed_id interprets char const* args as pointers, not strings.
This is probably not how it should be (not how I expect it to be, at least).
It goes wrong in:
- equals()
- hash() - fixed in the previous commit, but not good enough
Normally, there is no difference as there is only once instance for each possible
string value, but the current lambda call implementation introduces second instances.
The present commit fixes that immediate problem, but we may need a more general solution.